Revert "Optimize settings pages (#67)"
This reverts commit 78428182d89907af3f7ade17f32d49779a62ae38.
