+++ src/app/(main)/(admin)/settings/extensions/[identifier]/config/ConfigLoader.tsx
@@ -1,16 +1,15 @@
 'use client';
 
 import type { ConfigProps } from '@/app/(main)/(admin)/settings/extensions/[identifier]/config/common';
-import { ExtensionConfig, useExtensionConfig } from '@/components/extension-config';
+import { ExtensionConfig, useExtension, useExtensionConfig } from '@/components/extension-config';
 import { Alert, AlertDescription, AlertTitle } from '@/components/ui/alert';
 import { Button } from '@/components/ui/button';
 import { Dialog, DialogContent, DialogHeader, DialogTrigger } from '@/components/ui/dialog';
 import { Input } from '@/components/ui/input';
 import { ScrollArea } from '@/components/ui/scroll-area';
-import { useExtension } from '@/components/use-extension';
 import { rag } from '@/core/interface';
 import { getErrorMessage } from '@/lib/error';
-import { useState } from 'react';
+import { useState, useTransition } from 'react';
 
 export function ConfigLoader ({ base, identifier }: ConfigProps) {
   const Extension = useExtension(identifier);
@@ -18,7 +17,8 @@ export function ConfigLoader ({ base, identifier }: ConfigProps) {
   const hasConfig = Object.keys(Extension.optionsSchema.shape).length > 0;
 
   return (
-    <>
+    <div className="space-y-4">
+      <h1 className="text-xl font-semibold">{Extension.displayName} | Configuration</h1>
       {submitError != null && <Alert variant="destructive">
         <AlertTitle>{getErrorMessage(submitError)}</AlertTitle>
       </Alert>}
@@ -27,13 +27,13 @@ export function ConfigLoader ({ base, identifier }: ConfigProps) {
         {...props}
         onParsedValuesChange={() => {}}
       />}
-      {!hasConfig && <div className="text-sm">{Extension.displayName} has no custom options currently.</div>}
+      {!hasConfig && <div className='text-sm'>{Extension.displayName} has no custom options currently.</div>}
       <div className="flex gap-2 items-center">
         {hasConfig && <Button form={props.id} disabled={props.disabled}>Save</Button>}
         {hasConfig && <Button disabled={props.disabled} onClick={reset} variant="destructive">Reset</Button>}
         <Try options={props.values} identifier={identifier} />
       </div>
-    </>
+    </div>
   );
 }
