--- src/app/(main)/(admin)/settings/extensions/[identifier]/config/loading.tsx
@@ -1,9 +1,0 @@
-import { Loader } from '@/components/loader';
-
-export default function Loading () {
-  return (
-    <Loader loading>
-      Loading page...
-    </Loader>
-  );
-}
