+++ src/app/(main)/(admin)/settings/extensions/[identifier]/page.tsx
@@ -1,5 +1,4 @@
 import { getDef } from '@/app/(main)/(admin)/settings/extensions/utils';
-import { BackwardButton } from '@/components/navigation/backward-button';
 import { baseRegistry } from '@/rag-spec/base';
 import { notFound } from 'next/navigation';
 
@@ -20,10 +19,7 @@ export default async function Page ({ params }: { params: { identifier: string }
 
   return (
     <div className="p-4 space-y-4">
-      <h1 className="text-xl font-semibold">
-        <BackwardButton variant='ghost' />
-        {ctor.displayName}
-      </h1>
+      <h1 className="text-xl font-semibold">{ctor.displayName}</h1>
       {ctor.description && <article className="prose prose-sm prose-neutral dark:prose-invert overflow-x-hidden break-all">
         <ctor.description />
       </article>}
